Moving the schedule info under the games blueprint, enabling a more accurate games page
