+++ nbaspa_app/__init__.py
@@ -28,13 +28,11 @@ def create_app(config: str = "development"):
         from .games.routes import game_bp
         from .home.routes import home_bp
         from .players.routes import players_bp
-        from .schedule.routes import schedule_bp
         from .teams.routes import teams_bp
 
         app.register_blueprint(game_bp)
         app.register_blueprint(home_bp)
         app.register_blueprint(players_bp)
-        app.register_blueprint(schedule_bp)
         app.register_blueprint(teams_bp)
 
         return app
