+++ nbaspa_app/teams/routes.py
@@ -3,6 +3,7 @@
 from flask import Blueprint, render_template
 from flask import current_app as app
 from nbaspa.data.endpoints.parameters import SEASONS
+import pandas as pd
 
 from .data import (
     gen_teamlist,
@@ -64,6 +65,7 @@ def team_gamelog(teamid: int, season: int):
     teamlist = gen_teamlist(app=app)
     teamname = [row["teamname"] for row in teamlist if row["teamid"] == teamid][0]
     data = gen_gamelog(app=app, teamid=teamid, season=season)
+    data["GAME_DATE"] = pd.to_datetime(data["GAME_DATE"])
     data["W_PCT"] = (data["W_PCT"] * 100).round(1)
     return render_template(
         "gamelog.html",
