+++ nbaspa_app/templates/layout.html
@@ -34,7 +34,7 @@
                     <ul class="navbar-list">
                         <li class="navbar-item">
                             <a class="navbar-link" href="{{ url_for('home_bp.home') }}">Home</a>
-                            <a class="navbar-link" href="{{ url_for('schedule_bp.schedule') }}">Schedule</a>
+                            <a class="navbar-link" href="{{ url_for('game_bp.schedule') }}">Games</a>
                             <a class="navbar-link" href="{{ url_for('teams_bp.teams_home') }}">Teams</a>
                         </li>
                     </ul>
